Add firing_rates module
